Introduce delta protocols and alter the API
This is a major re-architecture and restructure of delta. And includes
some major changes:

## `DeltaItem` and `DeltaSection`
One of the major proof of the concepts hacks has been in the code base
too long—we used the hash value to determine if the data was the same
and equitability to determine if the data was equal. If the hash value
matched, but it wasn’t equal, we created a change record. This is
horrible, and is finally fixed. To determine it, we now require data to
confirm to the `DeltaItem` protocol, and  the `DeltaSection` protocol
for sectioned data. It requires the data to implement a
`deltaIdentifier` which we will use internally to determine changes.

## New API (public and internal)
Delta has a very broad and functional API, and we have multiple edge
cases. We therefor need a rather large public API that won’t really fit
into a single class. We previously tried to create a struct for each of
our edge cases, but the naming wasn’t obvious and the public API always
takes two arrays and return a single API, and you only wanted to
perform the calculations once, never any point to run it twice. The new
API is therefor just a set of functions with more obvious names.