Generating correct records when removing sections and moving items
When altering sections above a move item record, collection views and
table views expect the from index path’s section to be the original
section for the item, while the new index path should have the new
index path. This is slightly confusing in the context of how the other
APIs works.