Tweak API
After using the API in Flow Chat, I noticed the API could need some
tweaks to be even nicer. I expected `generateItemRecordsForSection` to
have a label for the section argument, considering it’s usually hard
coded. The function is now renamed to `generateItemRecords` and the
`section` label is now required. In these hardcoded cases the
`oldSection` will always be the same as the section, and I therefor
changed the `oldSection` argument to be optional, and set its default
value to be the already passed in section.