Remove change log from compile sources build phase
Xcode automatically adds markdown files to the “compile sources” build
phase when added to the project. This isn’t desirable as there are no
compiler for markdown files, and Xcode therefor generates a warning. I
don’t think it’s desirable to ship the change log with the framework
binary either.