Add preconditions to check for duplicate models
Delta assumes there are only one model with the same `deltaIdentifier`
when that assumption is violated, Delta has a tendency to crash. While
it's possible to work around it, it would complicate thing more then
necessary, especially as I consider it incorrect use and should be
treated as a bug from Delta's users.

When Delta crashes it's hard to understand that's the root of the
problem, I therefor added a `precondition` check that will fail if
duplicate models are passed in, giving a clear and concise error
message.