Fix exception when moving & reloading a cell
UITableView & UICollectionView do not allow moving & reloading a cell in the same batch of updates, doing so will cause an internal inconsistency exception.

To fix this, we queue the rows to be reloaded, and run another update block after all of the insertions / deletions / movements. When doing so, we now need to call reload on the new indexPath, since the item has already been moved.